backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
huggingface_hub
